Split the code concerning time into a dedicated module for future enhancement
